backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlmodel>=0.0.21
pydantic>=2.7.0
pydantic-settings>=2.4.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
psycopg2-binary>=2.9.9
gunicorn>=22.0.0
